requirements.txt
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pytest>=8.0.0
pytest-mock>=3.12.0
